+++ ZipKit/ZKDefs.m
@@ -20,19 +20,19 @@ NSString *const ZKFileAttributesKey = @"fileAttributes";
 NSString *const ZKPathKey = @"path";
 
 const unsigned long long ZKZipBlockSize = 262144;
-const NSUInteger ZKNotificationIterations = 100;
+const UInt32 ZKNotificationIterations = 100;
 
-const NSUInteger ZKCDHeaderMagicNumber = 0x02014B50;
-const NSUInteger ZKCDHeaderFixedDataLength = 46;
+const UInt32 ZKCDHeaderMagicNumber = 0x02014B50;
+const UInt32 ZKCDHeaderFixedDataLength = 46;
 
-const NSUInteger ZKCDTrailerMagicNumber = 0x06054B50;
-const NSUInteger ZKCDTrailerFixedDataLength = 22;
+const UInt32 ZKCDTrailerMagicNumber = 0x06054B50;
+const UInt32 ZKCDTrailerFixedDataLength = 22;
 
-const NSUInteger ZKLFHeaderMagicNumber = 0x04034B50;
-const NSUInteger ZKLFHeaderFixedDataLength = 30;
+const UInt32 ZKLFHeaderMagicNumber = 0x04034B50;
+const UInt32 ZKLFHeaderFixedDataLength = 30;
 
-const NSUInteger ZKCDTrailer64MagicNumber = 0x06064b50;
-const NSUInteger ZKCDTrailer64FixedDataLength = 56;
+const UInt32 ZKCDTrailer64MagicNumber = 0x06064b50;
+const UInt32 ZKCDTrailer64FixedDataLength = 56;
 
-const NSUInteger ZKCDTrailer64LocatorMagicNumber = 0x07064b50;
-const NSUInteger ZKCDTrailer64LocatorFixedDataLength = 20;
+const UInt32 ZKCDTrailer64LocatorMagicNumber = 0x07064b50;
+const UInt32 ZKCDTrailer64LocatorFixedDataLength = 20;
